python/pyproject.toml
[project]
name = "kaos-runtime"
version = "0.0.2.dev0"
description = "KAOS Runtime - K8s Agent Orchestration System"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "litellm>=1.0.0",
    "fastmcp>=1.0.0",
    "httpx>=0.25.0",
    "sse-starlette>=1.6.0",
    # OpenTelemetry core
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    # OpenTelemetry instrumentation
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    # OpenTelemetry exporters
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.0.0",
    "test-mcp-echo-server>=0.1.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "mcptools", "modelapi", "agent", "tests"]

[tool.black]
line-length = 100
target-version = ["py312"]
